Prevent exception when no input for update
Without this change, the UpdateOperator would throw an exception
stating that there was no valid page source. This occurs because the
driver which is responsible for setting the UpdateablePageSource
never calls the proper method due to never receiving any inputs.

This now handles the case where the page source is never set by
returning an EmptySplitPageSource